Time: 9 ms (54.59%) | Memory: 22.8 MB (74.85%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -25,7 +25,7 @@ public:
     {
         for (int i : p)
         {
-            h -= ceil((long double) i / speed);
+            h -= ceil((double) i / speed);
             if (h < 0)
                 return false;
         }
